Remove future package in favor of transaction package (#454)
