+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## What's Changed
 ### Breaking Changes
+* Remove `future` package.  Move package contents to `transaction`.
 * Remove `MakeLogicSigAccount` and replace with `MakeLogicSigAccountEscrow`. Mark `MakeLogicSig` as a private function as well, only intended for internal use.
 * Rename `SignLogicsigTransaction` to `SignLogicSigTransaction`.
 * Remove logicsig templates, `logic/langspec.json`, and all methods depending on it.
