+++ transaction/atomicTransactionComposer_test.go
@@ -1,4 +1,4 @@
-package future
+package transaction
 
 import (
 	"testing"
