+++ transaction/transaction_test.go
@@ -4,12 +4,11 @@ import (
 	"encoding/base64"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/algorand/go-algorand-sdk/crypto"
 	"github.com/algorand/go-algorand-sdk/encoding/msgpack"
 	"github.com/algorand/go-algorand-sdk/mnemonic"
 	"github.com/algorand/go-algorand-sdk/types"
+	"github.com/stretchr/testify/require"
 )
 
 func byteFromBase64(s string) []byte {
@@ -26,6 +25,15 @@ func byte32ArrayFromBase64(s string) (out [32]byte) {
 	return
 }
 
+func byte64ArrayFromBase64(s string) (out [64]byte) {
+	slice := byteFromBase64(s)
+	if len(slice) != 64 {
+		panic("wrong length: input slice not 64 bytes")
+	}
+	copy(out[:], slice)
+	return
+}
+
 func TestMakePaymentTxn(t *testing.T) {
 	const fromAddress = "47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU"
 	const toAddress = "PNWOET7LLOWMBMLE4KOCELCX6X3D3Q4H2Q4QJASYIEOF7YIPPQBG3YQ5YI"
@@ -34,13 +42,21 @@ func TestMakePaymentTxn(t *testing.T) {
 	const golden = "gqNzaWfEQPhUAZ3xkDDcc8FvOVo6UinzmKBCqs0woYSfodlmBMfQvGbeUx3Srxy3dyJDzv7rLm26BRv9FnL2/AuT7NYfiAWjdHhui6NhbXTNA+ilY2xvc2XEIEDpNJKIJWTLzpxZpptnVCaJ6aHDoqnqW2Wm6KRCH/xXo2ZlZc0EmKJmds0wsqNnZW6sZGV2bmV0LXYzMy4womdoxCAmCyAJoJOohot5WHIvpeVG7eftF+TYXEx4r7BFJpDt0qJsds00mqRub3RlxAjqABVHQ2y/lqNyY3bEIHts4k/rW6zAsWTinCIsV/X2PcOH1DkEglhBHF/hD3wCo3NuZMQg5/D4TQaBHfnzHI2HixFV9GcdUaGFwgCQhmf0SVhwaKGkdHlwZaNwYXk="
 	gh := byteFromBase64("JgsgCaCTqIaLeVhyL6XlRu3n7Rfk2FxMeK+wRSaQ7dI=")
 
-	txn, err := MakePaymentTxn(fromAddress, toAddress, 4, 1000, 12466, 13466, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", "devnet-v33.0", gh)
+	params := types.SuggestedParams{
+		Fee:             4,
+		FirstRoundValid: 12466,
+		LastRoundValid:  13466,
+		GenesisID:       "devnet-v33.0",
+		GenesisHash:     gh,
+	}
+	txn, err := MakePaymentTxn(fromAddress, toAddress, 1000, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", params)
 	require.NoError(t, err)
 
 	key, err := mnemonic.ToPrivateKey(mn)
 	require.NoError(t, err)
 
 	id, bytes, err := crypto.SignTransaction(key, txn)
+	require.NoError(t, err)
 
 	stxBytes := byteFromBase64(golden)
 	require.Equal(t, stxBytes, bytes)
@@ -52,10 +68,14 @@ func TestMakePaymentTxn(t *testing.T) {
 func TestMakePaymentTxn2(t *testing.T) {
 	const fromAddress = "47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU"
 	const toAddress = "PNWOET7LLOWMBMLE4KOCELCX6X3D3Q4H2Q4QJASYIEOF7YIPPQBG3YQ5YI"
-
-	_, err := MakePaymentTxn(fromAddress, toAddress, 4, 1000, 12466, 13466, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", "devnet-v33.0", []byte{})
+	params := types.SuggestedParams{
+		Fee:             4,
+		FirstRoundValid: 12466,
+		LastRoundValid:  13466,
+		GenesisID:       "devnet-v33.0",
+	}
+	_, err := MakePaymentTxn(fromAddress, toAddress, 1000, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", params)
 	require.Error(t, err)
-
 }
 
 func TestMakePaymentTxnWithLease(t *testing.T) {
@@ -65,9 +85,15 @@ func TestMakePaymentTxnWithLease(t *testing.T) {
 	const mn = "advice pudding treat near rule blouse same whisper inner electric quit surface sunny dismiss leader blood seat clown cost exist hospital century reform able sponsor"
 	const golden = "gqNzaWfEQOMmFSIKsZvpW0txwzhmbgQjxv6IyN7BbV5sZ2aNgFbVcrWUnqPpQQxfPhV/wdu9jzEPUU1jAujYtcNCxJ7ONgejdHhujKNhbXTNA+ilY2xvc2XEIEDpNJKIJWTLzpxZpptnVCaJ6aHDoqnqW2Wm6KRCH/xXo2ZlZc0FLKJmds0wsqNnZW6sZGV2bmV0LXYzMy4womdoxCAmCyAJoJOohot5WHIvpeVG7eftF+TYXEx4r7BFJpDt0qJsds00mqJseMQgAQIDBAECAwQBAgMEAQIDBAECAwQBAgMEAQIDBAECAwSkbm90ZcQI6gAVR0Nsv5ajcmN2xCB7bOJP61uswLFk4pwiLFf19j3Dh9Q5BIJYQRxf4Q98AqNzbmTEIOfw+E0GgR358xyNh4sRVfRnHVGhhcIAkIZn9ElYcGihpHR5cGWjcGF5"
 	gh := byteFromBase64("JgsgCaCTqIaLeVhyL6XlRu3n7Rfk2FxMeK+wRSaQ7dI=")
-
+	params := types.SuggestedParams{
+		Fee:             4,
+		FirstRoundValid: 12466,
+		LastRoundValid:  13466,
+		GenesisID:       "devnet-v33.0",
+		GenesisHash:     gh,
+	}
 	lease := [32]byte{1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4}
-	txn, err := MakePaymentTxn(fromAddress, toAddress, 4, 1000, 12466, 13466, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", "devnet-v33.0", gh)
+	txn, err := MakePaymentTxn(fromAddress, toAddress, 1000, byteFromBase64("6gAVR0Nsv5Y="), "IDUTJEUIEVSMXTU4LGTJWZ2UE2E6TIODUKU6UW3FU3UKIQQ77RLUBBBFLA", params)
 	require.NoError(t, err)
 	txn.AddLease(lease, 4)
 	require.NoError(t, err)
@@ -76,6 +102,7 @@ func TestMakePaymentTxnWithLease(t *testing.T) {
 	require.NoError(t, err)
 
 	id, stxBytes, err := crypto.SignTransaction(key, txn)
+	require.NoError(t, err)
 
 	goldenBytes := byteFromBase64(golden)
 	require.Equal(t, goldenBytes, stxBytes)
@@ -118,8 +145,14 @@ func TestKeyRegTxn(t *testing.T) {
 
 func TestMakeKeyRegTxn(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
-	tx, err := MakeKeyRegTxn(addr, 10, 322575, 323575, []byte{45, 67}, "", "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=",
-		"Kv7QI7chi1y6axoy+t7wzAVpePqRq/rkjzWh/RMYyLo=", "bPgrv4YogPcdaUAxrt1QysYZTVyRAuUMD4zQmCu9llc=", 10000, 10111, 11)
+	ghAsArray := byte32ArrayFromBase64("SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=")
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeKeyRegTxn(addr, []byte{45, 67}, params, "Kv7QI7chi1y6axoy+t7wzAVpePqRq/rkjzWh/RMYyLo=", "bPgrv4YogPcdaUAxrt1QysYZTVyRAuUMD4zQmCu9llc=", 10000, 10111, 11)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(addr)
@@ -146,19 +179,88 @@ func TestMakeKeyRegTxn(t *testing.T) {
 	require.Equal(t, expKeyRegTxn, tx)
 }
 
+func TestMakeKeyRegTxnWithStateProofKey(t *testing.T) {
+	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
+	ghAsArray := byte32ArrayFromBase64("SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=")
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	// nonparticipation
+	tx, err := MakeKeyRegTxnWithStateProofKey(addr, []byte{45, 67}, params, "", "", "", 0, 0, 0, true)
+	require.NoError(t, err)
+	a, err := types.DecodeAddress(addr)
+	require.NoError(t, err)
+	expKeyRegTxn := types.Transaction{
+		Type: types.KeyRegistrationTx,
+		Header: types.Header{
+			Sender:      a,
+			Fee:         2020,
+			FirstValid:  322575,
+			LastValid:   323575,
+			Note:        []byte{45, 67},
+			GenesisHash: byte32ArrayFromBase64("SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="),
+			GenesisID:   "",
+		},
+		KeyregTxnFields: types.KeyregTxnFields{
+			Nonparticipation: true,
+		},
+	}
+	require.Equal(t, expKeyRegTxn, tx)
+
+	// online
+	stateProof := "mYR0GVEObMTSNdsKM6RwYywHYPqVDqg3E4JFzxZOreH9NU8B+tKzUanyY8AQ144hETgSMX7fXWwjBdHz6AWk9w=="
+	tx, err = MakeKeyRegTxnWithStateProofKey(addr, []byte{45, 67}, params, "Kv7QI7chi1y6axoy+t7wzAVpePqRq/rkjzWh/RMYyLo=", "bPgrv4YogPcdaUAxrt1QysYZTVyRAuUMD4zQmCu9llc=", stateProof, 10000, 10111, 11, false)
+	require.NoError(t, err)
+
+	a, err = types.DecodeAddress(addr)
+	require.NoError(t, err)
+	expKeyRegTxn = types.Transaction{
+		Type: types.KeyRegistrationTx,
+		Header: types.Header{
+			Sender:      a,
+			Fee:         3800,
+			FirstValid:  322575,
+			LastValid:   323575,
+			Note:        []byte{45, 67},
+			GenesisHash: byte32ArrayFromBase64("SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="),
+			GenesisID:   "",
+		},
+		KeyregTxnFields: types.KeyregTxnFields{
+			VotePK:           byte32ArrayFromBase64("Kv7QI7chi1y6axoy+t7wzAVpePqRq/rkjzWh/RMYyLo="),
+			SelectionPK:      byte32ArrayFromBase64("bPgrv4YogPcdaUAxrt1QysYZTVyRAuUMD4zQmCu9llc="),
+			VoteFirst:        10000,
+			VoteLast:         10111,
+			VoteKeyDilution:  11,
+			Nonparticipation: false,
+			StateProofPK:     byte64ArrayFromBase64(stateProof),
+		},
+	}
+	require.Equal(t, expKeyRegTxn, tx)
+}
+
 func TestMakeAssetCreateTxn(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const defaultFrozen = false
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const total = 100
 	const reserve = addr
 	const freeze = addr
 	const clawback = addr
 	const unitName = "tst"
 	const assetName = "testcoin"
-	const testURL = "websitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsitewebsi" //96 characters
+	const testURL = "website"
 	const metadataHash = "fACPO4nRgO55j1ndAK3W6Sgc4APkcyFh"
-	tx, err := MakeAssetCreateTxn(addr, 10, 322575, 323575, nil, "", genesisHash, total, 0, defaultFrozen, addr, reserve, freeze, clawback, unitName, assetName, testURL, metadataHash)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetCreateTxn(addr, nil, params, total, 0, defaultFrozen, addr, reserve, freeze, clawback, unitName, assetName, testURL, metadataHash)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(addr)
@@ -167,7 +269,7 @@ func TestMakeAssetCreateTxn(t *testing.T) {
 		Type: types.AssetConfigTx,
 		Header: types.Header{
 			Sender:      a,
-			Fee:         4920,
+			Fee:         4020,
 			FirstValid:  322575,
 			LastValid:   323575,
 			GenesisHash: byte32ArrayFromBase64(genesisHash),
@@ -192,7 +294,8 @@ func TestMakeAssetCreateTxn(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
-	signedGolden := "gqNzaWfEQCPNXED15sdMwiSxl9K4LALy5brxBpnOtBurPK2TAtjXpxKQqN9flVtjxP0GGYVUjNx/Sez0W4v1CSU9FhYSUgOjdHhuh6RhcGFyiaJhbcQgZkFDUE80blJnTzU1ajFuZEFLM1c2U2djNEFQa2N5RmiiYW6odGVzdGNvaW6iYXXZYHdlYnNpdGV3ZWJzaXRld2Vic2l0ZXdlYnNpdGV3ZWJzaXRld2Vic2l0ZXdlYnNpdGV3ZWJzaXRld2Vic2l0ZXdlYnNpdGV3ZWJzaXRld2Vic2l0ZXdlYnNpdGV3ZWJzaaFjxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFmxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFtxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFyxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aF0ZKJ1bqN0c3SjZmVlzRM4omZ2zgAE7A+iZ2jEIEhjtRiks8hOyBDyLU8QgcsPcfBZp6wg3sYvf3DlCToiomx2zgAE7/ejc25kxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aR0eXBlpGFjZmc="
+	require.NoError(t, err)
+	signedGolden := "gqNzaWfEQEDd1OMRoQI/rzNlU4iiF50XQXmup3k5czI9hEsNqHT7K4KsfmA/0DUVkbzOwtJdRsHS8trm3Arjpy9r7AXlbAujdHhuh6RhcGFyiaJhbcQgZkFDUE80blJnTzU1ajFuZEFLM1c2U2djNEFQa2N5RmiiYW6odGVzdGNvaW6iYXWnd2Vic2l0ZaFjxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFmxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFtxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFyxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aF0ZKJ1bqN0c3SjZmVlzQ+0omZ2zgAE7A+iZ2jEIEhjtRiks8hOyBDyLU8QgcsPcfBZp6wg3sYvf3DlCToiomx2zgAE7/ejc25kxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aR0eXBlpGFjZmc="
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
 
@@ -200,6 +303,7 @@ func TestMakeAssetCreateTxnWithDecimals(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const defaultFrozen = false
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const total = 100
 	const decimals = 1
 	const reserve = addr
@@ -209,7 +313,13 @@ func TestMakeAssetCreateTxnWithDecimals(t *testing.T) {
 	const assetName = "testcoin"
 	const testURL = "website"
 	const metadataHash = "fACPO4nRgO55j1ndAK3W6Sgc4APkcyFh"
-	tx, err := MakeAssetCreateTxn(addr, 10, 322575, 323575, nil, "", genesisHash, total, decimals, defaultFrozen, addr, reserve, freeze, clawback, unitName, assetName, testURL, metadataHash)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetCreateTxn(addr, nil, params, total, decimals, defaultFrozen, addr, reserve, freeze, clawback, unitName, assetName, testURL, metadataHash)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(addr)
@@ -244,6 +354,7 @@ func TestMakeAssetCreateTxnWithDecimals(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQCj5xLqNozR5ahB+LNBlTG+d0gl0vWBrGdAXj1ibsCkvAwOsXs5KHZK1YdLgkdJecQiWm4oiZ+pm5Yg0m3KFqgqjdHhuh6RhcGFyiqJhbcQgZkFDUE80blJnTzU1ajFuZEFLM1c2U2djNEFQa2N5RmiiYW6odGVzdGNvaW6iYXWnd2Vic2l0ZaFjxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aJkYwGhZsQgCfvSdiwI+Gxa5r9t16epAd5mdddQ4H6MXHaYZH224f2hbcQgCfvSdiwI+Gxa5r9t16epAd5mdddQ4H6MXHaYZH224f2hcsQgCfvSdiwI+Gxa5r9t16epAd5mdddQ4H6MXHaYZH224f2hdGSidW6jdHN0o2ZlZc0P3KJmds4ABOwPomdoxCBIY7UYpLPITsgQ8i1PEIHLD3HwWaesIN7GL39w5Qk6IqJsds4ABO/3o3NuZMQgCfvSdiwI+Gxa5r9t16epAd5mdddQ4H6MXHaYZH224f2kdHlwZaRhY2Zn"
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
@@ -251,13 +362,19 @@ func TestMakeAssetCreateTxnWithDecimals(t *testing.T) {
 func TestMakeAssetConfigTxn(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const manager = addr
 	const reserve = addr
 	const freeze = addr
 	const clawback = addr
 	const assetIndex = 1234
-	tx, err := MakeAssetConfigTxn(addr, 10, 322575, 323575, nil, "", genesisHash,
-		assetIndex, manager, reserve, freeze, clawback, false)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetConfigTxn(addr, nil, params, assetIndex, manager, reserve, freeze, clawback, false)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(addr)
@@ -287,6 +404,7 @@ func TestMakeAssetConfigTxn(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQBBkfw5n6UevuIMDo2lHyU4dS80JCCQ/vTRUcTx5m0ivX68zTKyuVRrHaTbxbRRc3YpJ4zeVEnC9Fiw3Wf4REwejdHhuiKRhcGFyhKFjxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFmxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFtxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aFyxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aRjYWlkzQTSo2ZlZc0NSKJmds4ABOwPomdoxCBIY7UYpLPITsgQ8i1PEIHLD3HwWaesIN7GL39w5Qk6IqJsds4ABO/3o3NuZMQgCfvSdiwI+Gxa5r9t16epAd5mdddQ4H6MXHaYZH224f2kdHlwZaRhY2Zn"
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
@@ -294,24 +412,37 @@ func TestMakeAssetConfigTxn(t *testing.T) {
 func TestMakeAssetConfigTxnStrictChecking(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const manager = addr
 	const reserve = addr
 	const freeze = ""
 	const clawback = addr
 	const assetIndex = 1234
-	_, err := MakeAssetConfigTxn(addr, 10, 322575, 323575, nil, "", genesisHash,
-		assetIndex, manager, reserve, freeze, clawback, true)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: 322575,
+		LastRoundValid:  323575,
+		GenesisHash:     ghAsArray[:],
+	}
+	_, err := MakeAssetConfigTxn(addr, nil, params, assetIndex, manager, reserve, freeze, clawback, true)
 	require.Error(t, err)
 }
 
 func TestMakeAssetDestroyTxn(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const creator = addr
 	const assetIndex = 1
 	const firstValidRound = 322575
 	const lastValidRound = 323575
-	tx, err := MakeAssetDestroyTxn(creator, 10, firstValidRound, lastValidRound, nil, "", genesisHash, assetIndex)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: firstValidRound,
+		LastRoundValid:  lastValidRound,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetDestroyTxn(creator, nil, params, assetIndex)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(creator)
@@ -336,6 +467,7 @@ func TestMakeAssetDestroyTxn(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQBSP7HtzD/Lvn4aVvaNpeR4T93dQgo4LvywEwcZgDEoc/WVl3aKsZGcZkcRFoiWk8AidhfOZzZYutckkccB8RgGjdHhuh6RjYWlkAaNmZWXNB1iiZnbOAATsD6JnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHbOAATv96NzbmTEIAn70nYsCPhsWua/bdenqQHeZnXXUOB+jFx2mGR9tuH9pHR5cGWkYWNmZw=="
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
@@ -343,12 +475,19 @@ func TestMakeAssetDestroyTxn(t *testing.T) {
 func TestMakeAssetFreezeTxn(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const assetIndex = 1
 	const firstValidRound = 322575
 	const lastValidRound = 323576
 	const freezeSetting = true
 	const target = addr
-	tx, err := MakeAssetFreezeTxn(addr, 10, firstValidRound, lastValidRound, nil, "", genesisHash, assetIndex, target, freezeSetting)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: firstValidRound,
+		LastRoundValid:  lastValidRound,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetFreezeTxn(addr, nil, params, assetIndex, target, freezeSetting)
 	require.NoError(t, err)
 
 	a, err := types.DecodeAddress(addr)
@@ -374,6 +513,7 @@ func TestMakeAssetFreezeTxn(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQAhru5V2Xvr19s4pGnI0aslqwY4lA2skzpYtDTAN9DKSH5+qsfQQhm4oq+9VHVj7e1rQC49S28vQZmzDTVnYDQGjdHhuiaRhZnJ6w6RmYWRkxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aRmYWlkAaNmZWXNCRqiZnbOAATsD6JnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHbOAATv+KNzbmTEIAn70nYsCPhsWua/bdenqQHeZnXXUOB+jFx2mGR9tuH9pHR5cGWkYWZyeg=="
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
@@ -385,14 +525,19 @@ func TestMakeAssetTransferTxn(t *testing.T) {
 
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const sender, recipient, closeAssetsTo = addr, addr, addr
 	const assetIndex = 1
 	const firstValidRound = 322575
 	const lastValidRound = 323576
 	const amountToSend = 1
-
-	tx, err := MakeAssetTransferTxn(sender, recipient, closeAssetsTo, amountToSend, 10, firstValidRound,
-		lastValidRound, nil, "", genesisHash, assetIndex)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: firstValidRound,
+		LastRoundValid:  lastValidRound,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetTransferTxn(sender, recipient, amountToSend, nil, params, closeAssetsTo, assetIndex)
 	require.NoError(t, err)
 
 	sendAddr, err := types.DecodeAddress(sender)
@@ -435,12 +580,17 @@ func TestMakeAssetTransferTxn(t *testing.T) {
 func TestMakeAssetAcceptanceTxn(t *testing.T) {
 	const sender = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const assetIndex = 1
 	const firstValidRound = 322575
 	const lastValidRound = 323575
-
-	tx, err := MakeAssetAcceptanceTxn(sender, 10, firstValidRound,
-		lastValidRound, nil, "", genesisHash, assetIndex)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: firstValidRound,
+		LastRoundValid:  lastValidRound,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetAcceptanceTxn(sender, nil, params, assetIndex)
 	require.NoError(t, err)
 
 	sendAddr, err := types.DecodeAddress(sender)
@@ -469,6 +619,7 @@ func TestMakeAssetAcceptanceTxn(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQJ7q2rOT8Sb/wB0F87ld+1zMprxVlYqbUbe+oz0WM63FctIi+K9eYFSqT26XBZ4Rr3+VTJpBE+JLKs8nctl9hgijdHhuiKRhcmN2xCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aNmZWXNCOiiZnbOAATsD6JnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHbOAATv96NzbmTEIAn70nYsCPhsWua/bdenqQHeZnXXUOB+jFx2mGR9tuH9pHR5cGWlYXhmZXKkeGFpZAE="
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
 }
@@ -476,14 +627,20 @@ func TestMakeAssetAcceptanceTxn(t *testing.T) {
 func TestMakeAssetRevocationTransaction(t *testing.T) {
 	const addr = "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4"
 	const genesisHash = "SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI="
+	ghAsArray := byte32ArrayFromBase64(genesisHash)
 	const revoker, recipient, revoked = addr, addr, addr
 	const assetIndex = 1
 	const firstValidRound = 322575
 	const lastValidRound = 323575
 	const amountToSend = 1
 
-	tx, err := MakeAssetRevocationTxn(revoker, revoked, recipient, amountToSend, 10, firstValidRound,
-		lastValidRound, nil, "", genesisHash, assetIndex)
+	params := types.SuggestedParams{
+		Fee:             10,
+		FirstRoundValid: firstValidRound,
+		LastRoundValid:  lastValidRound,
+		GenesisHash:     ghAsArray[:],
+	}
+	tx, err := MakeAssetRevocationTxn(revoker, revoked, amountToSend, recipient, nil, params, assetIndex)
 	require.NoError(t, err)
 
 	sendAddr, err := types.DecodeAddress(revoker)
@@ -520,8 +677,105 @@ func TestMakeAssetRevocationTransaction(t *testing.T) {
 	private, err := mnemonic.ToPrivateKey(addrSK)
 	require.NoError(t, err)
 	_, newStxBytes, err := crypto.SignTransaction(private, tx)
+	require.NoError(t, err)
 	signedGolden := "gqNzaWfEQHsgfEAmEHUxLLLR9s+Y/yq5WeoGo/jAArCbany+7ZYwExMySzAhmV7M7S8+LBtJalB4EhzEUMKmt3kNKk6+vAWjdHhuiqRhYW10AaRhcmN2xCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aRhc25kxCAJ+9J2LAj4bFrmv23Xp6kB3mZ111Dgfoxcdphkfbbh/aNmZWXNCqqiZnbOAATsD6JnaMQgSGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiKibHbOAATv96NzbmTEIAn70nYsCPhsWua/bdenqQHeZnXXUOB+jFx2mGR9tuH9pHR5cGWlYXhmZXKkeGFpZAE="
 	require.EqualValues(t, newStxBytes, byteFromBase64(signedGolden))
+}
+
+func TestMakeApplicationCallTx(t *testing.T) {
+	const fee = 1000
+	const firstRound = 2063137
+	const genesisID = "devnet-v1.0"
+	genesisHash := byteFromBase64("sC3P7e2SdbqKJK0tbiCdK9tdSpbe6XeCGKdoNzmlj0E=")
+
+	params := types.SuggestedParams{
+		Fee:             fee,
+		FirstRoundValid: firstRound,
+		LastRoundValid:  firstRound + 1000,
+		GenesisHash:     genesisHash,
+		GenesisID:       genesisID,
+		FlatFee:         true,
+	}
+	note := byteFromBase64("8xMCTuLQ810=")
+	program := []byte{1, 32, 1, 1, 34}
+	args := make([][]byte, 2)
+	args[0] = []byte("123")
+	args[1] = []byte("456")
+	foreignApps := make([]uint64, 1)
+	foreignApps[0] = 10
+	foreignAssets := foreignApps
+	gSchema := types.StateSchema{NumUint: uint64(1), NumByteSlice: uint64(1)}
+	lSchema := types.StateSchema{NumUint: uint64(1), NumByteSlice: uint64(1)}
+	extraPages := uint32(2)
+	addr := make([]string, 1)
+	addr[0] = "47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU"
+	boxReferences := make([]types.AppBoxReference, 3)
+	boxReferences[0] = types.AppBoxReference{AppID: 2, Name: []byte("box_name")}
+	boxReferences[1] = types.AppBoxReference{AppID: 10, Name: []byte("box_name")}
+	boxReferences[2] = types.AppBoxReference{AppID: 10, Name: []byte("box_name2")}
+
+	tx, err := MakeApplicationCallTxWithBoxes(2, args, addr, foreignApps, foreignAssets, boxReferences, types.NoOpOC, program, program, gSchema, lSchema, extraPages, params, types.Address{}, note, types.Digest{}, [32]byte{}, types.Address{})
+	require.NoError(t, err)
+	require.EqualValues(t, 2, tx.ExtraProgramPages)
+
+	// verify that the correct app index was calculated
+	require.EqualValues(t, 0, tx.BoxReferences[0].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[1].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[2].ForeignAppIdx)
+
+	// the current app can also be referenced with AppID = 0
+	boxReferences[0].AppID = 0
+	tx, err = MakeApplicationCallTxWithBoxes(2, args, addr, foreignApps, foreignAssets, boxReferences, types.NoOpOC, program, program, gSchema, lSchema, extraPages, params, types.Address{}, note, types.Digest{}, [32]byte{}, types.Address{})
+	require.NoError(t, err)
+	require.EqualValues(t, 0, tx.BoxReferences[0].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[1].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[2].ForeignAppIdx)
+
+	// if the current app's ID is provided explicitly AND is present in the foreignApps array
+	// then the index in the array should be returned rather than the usual value of 0
+	boxReferences[0].AppID = 2
+	foreignApps = append(foreignApps, 2)
+	tx, err = MakeApplicationCallTxWithBoxes(2, args, addr, foreignApps, foreignAssets, boxReferences, types.NoOpOC, program, program, gSchema, lSchema, extraPages, params, types.Address{}, note, types.Digest{}, [32]byte{}, types.Address{})
+	require.NoError(t, err)
+	require.EqualValues(t, 2, tx.BoxReferences[0].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[1].ForeignAppIdx)
+	require.EqualValues(t, 1, tx.BoxReferences[2].ForeignAppIdx)
+}
+
+func TestMakeApplicationCallTxInvalidBoxes(t *testing.T) {
+	const fee = 1000
+	const firstRound = 2063137
+	const genesisID = "devnet-v1.0"
+	genesisHash := byteFromBase64("sC3P7e2SdbqKJK0tbiCdK9tdSpbe6XeCGKdoNzmlj0E=")
+
+	params := types.SuggestedParams{
+		Fee:             fee,
+		FirstRoundValid: firstRound,
+		LastRoundValid:  firstRound + 1000,
+		GenesisHash:     genesisHash,
+		GenesisID:       genesisID,
+		FlatFee:         true,
+	}
+	note := byteFromBase64("8xMCTuLQ810=")
+	program := []byte{1, 32, 1, 1, 34}
+	args := make([][]byte, 2)
+	args[0] = []byte("123")
+	args[1] = []byte("456")
+	foreignApps := make([]uint64, 1)
+	foreignApps[0] = 10
+	foreignAssets := foreignApps
+	gSchema := types.StateSchema{NumUint: uint64(1), NumByteSlice: uint64(1)}
+	lSchema := types.StateSchema{NumUint: uint64(1), NumByteSlice: uint64(1)}
+	extraPages := uint32(2)
+	addr := make([]string, 1)
+	addr[0] = "47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU"
+	boxReferences := make([]types.AppBoxReference, 3)
+	boxReferences[0] = types.AppBoxReference{AppID: 2, Name: []byte("box_name")}
+	boxReferences[1] = types.AppBoxReference{AppID: 10, Name: []byte("box_name")}
+	boxReferences[2] = types.AppBoxReference{AppID: 11, Name: []byte("box_name")}
+
+	_, err := MakeApplicationCallTxWithBoxes(2, args, addr, foreignApps, foreignAssets, boxReferences, types.NoOpOC, program, program, gSchema, lSchema, extraPages, params, types.Address{}, note, types.Digest{}, [32]byte{}, types.Address{})
+	require.Error(t, err, "the app id 10 provided for this box is not in the foreignApps array")
 }
 
 func TestComputeGroupID(t *testing.T) {
@@ -535,19 +789,29 @@ func TestComputeGroupID(t *testing.T) {
 	genesisHash := byteFromBase64("sC3P7e2SdbqKJK0tbiCdK9tdSpbe6XeCGKdoNzmlj0E=")
 
 	const firstRound1 = 710399
+	params1 := types.SuggestedParams{
+		Fee:             fee,
+		FirstRoundValid: firstRound1,
+		LastRoundValid:  firstRound1 + 1000,
+		GenesisHash:     genesisHash,
+		GenesisID:       genesisID,
+		FlatFee:         true,
+	}
 	note1 := byteFromBase64("wRKw5cJ0CMo=")
-	tx1, err := MakePaymentTxnWithFlatFee(
-		fromAddress, toAddress, fee, amount, firstRound1, firstRound1+1000,
-		note1, "", genesisID, genesisHash,
-	)
+	tx1, err := MakePaymentTxn(fromAddress, toAddress, amount, note1, "", params1)
 	require.NoError(t, err)
 
 	const firstRound2 = 710515
+	params2 := types.SuggestedParams{
+		Fee:             fee,
+		FirstRoundValid: firstRound2,
+		LastRoundValid:  firstRound2 + 1000,
+		GenesisHash:     genesisHash,
+		GenesisID:       genesisID,
+		FlatFee:         true,
+	}
 	note2 := byteFromBase64("dBlHI6BdrIg=")
-	tx2, err := MakePaymentTxnWithFlatFee(
-		fromAddress, toAddress, fee, amount, firstRound2, firstRound2+1000,
-		note2, "", genesisID, genesisHash,
-	)
+	tx2, err := MakePaymentTxn(fromAddress, toAddress, amount, note2, "", params2)
 	require.NoError(t, err)
 
 	const goldenTx1 = "gaN0eG6Ko2FtdM0H0KNmZWXNA+iiZnbOAArW/6NnZW6rZGV2bmV0LXYxLjCiZ2jEILAtz+3tknW6iiStLW4gnSvbXUqW3ul3ghinaDc5pY9Bomx2zgAK2uekbm90ZcQIwRKw5cJ0CMqjcmN2xCCj8AKs8kPYlx63ppj1w5410qkMRGZ9FYofNYPXxGpNLKNzbmTEIKPwAqzyQ9iXHremmPXDnjXSqQxEZn0Vih81g9fEak0spHR5cGWjcGF5"
@@ -560,6 +824,7 @@ func TestComputeGroupID(t *testing.T) {
 	require.Equal(t, byteFromBase64(goldenTx2), msgpack.Encode(stx2))
 
 	gid, err := crypto.ComputeGroupID([]types.Transaction{tx1, tx2})
+	require.NoError(t, err)
 
 	// goal clerk group sets Group to every transaction and concatenate them in output file
 	// simulating that behavior here
@@ -574,7 +839,7 @@ func TestComputeGroupID(t *testing.T) {
 
 	require.Equal(t, byteFromBase64(goldenTxg), txg)
 
-	// check AssignGroupID, do not validate correctness of Group field calculation
+	// check transaction.AssignGroupID, do not validate correctness of Group field calculation
 	result, err := AssignGroupID([]types.Transaction{tx1, tx2}, "BH55E5RMBD4GYWXGX5W5PJ5JAHPGM5OXKDQH5DC4O2MGI7NW4H6VOE4CP4")
 	require.NoError(t, err)
 	require.Equal(t, 0, len(result))
@@ -586,4 +851,207 @@ func TestComputeGroupID(t *testing.T) {
 	result, err = AssignGroupID([]types.Transaction{tx1, tx2}, "")
 	require.NoError(t, err)
 	require.Equal(t, 2, len(result))
+}
+
+func TestLogicSig(t *testing.T) {
+	// validate LogicSig signed transaction against goal
+	const fromAddress = "47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU"
+	const toAddress = "PNWOET7LLOWMBMLE4KOCELCX6X3D3Q4H2Q4QJASYIEOF7YIPPQBG3YQ5YI"
+	const mn = "advice pudding treat near rule blouse same whisper inner electric quit surface sunny dismiss leader blood seat clown cost exist hospital century reform able sponsor"
+	const fee = 1000
+	const amount = 2000
+	const firstRound = 2063137
+	const genesisID = "devnet-v1.0"
+	genesisHash := byteFromBase64("sC3P7e2SdbqKJK0tbiCdK9tdSpbe6XeCGKdoNzmlj0E=")
+	note := byteFromBase64("8xMCTuLQ810=")
+
+	params := types.SuggestedParams{
+		Fee:             fee,
+		FirstRoundValid: firstRound,
+		LastRoundValid:  firstRound + 1000,
+		GenesisHash:     genesisHash,
+		GenesisID:       genesisID,
+		FlatFee:         true,
+	}
+	tx, err := MakePaymentTxn(fromAddress, toAddress, amount, note, "", params)
+	require.NoError(t, err)
+
+	// goal clerk send -o tx3 -a 2000 --fee 1000 -d ~/.algorand -w test -L sig.lsig --argb64 MTIz --argb64 NDU2 \
+	// -f 47YPQTIGQEO7T4Y4RWDYWEKV6RTR2UNBQXBABEEGM72ESWDQNCQ52OPASU \
+	// -t PNWOET7LLOWMBMLE4KOCELCX6X3D3Q4H2Q4QJASYIEOF7YIPPQBG3YQ5YI
+	const golden = "gqRsc2lng6NhcmeSxAMxMjPEAzQ1NqFsxAUBIAEBIqNzaWfEQE6HXaI5K0lcq50o/y3bWOYsyw9TLi/oorZB4xaNdn1Z14351u2f6JTON478fl+JhIP4HNRRAIh/I8EWXBPpJQ2jdHhuiqNhbXTNB9CjZmVlzQPoomZ2zgAfeyGjZ2Vuq2Rldm5ldC12MS4womdoxCCwLc/t7ZJ1uookrS1uIJ0r211Klt7pd4IYp2g3OaWPQaJsds4AH38JpG5vdGXECPMTAk7i0PNdo3JjdsQge2ziT+tbrMCxZOKcIixX9fY9w4fUOQSCWEEcX+EPfAKjc25kxCDn8PhNBoEd+fMcjYeLEVX0Zx1RoYXCAJCGZ/RJWHBooaR0eXBlo3BheQ=="
+
+	program := []byte{1, 32, 1, 1, 34}
+	args := make([][]byte, 2)
+	args[0] = []byte("123")
+	args[1] = []byte("456")
+	key, err := mnemonic.ToPrivateKey(mn)
+	require.NoError(t, err)
+	lsig, err := crypto.MakeLogicSigAccountDelegated(program, args, key)
+	require.NoError(t, err)
+
+	_, stxBytes, err := crypto.SignLogicSigAccountTransaction(lsig, tx)
+	require.NoError(t, err)
+
+	require.Equal(t, byteFromBase64(golden), stxBytes)
+}
+
+func TestFee(t *testing.T) {
+	testcases := []struct {
+		name     string
+		flatFee  bool
+		fee      types.MicroAlgos
+		expected types.MicroAlgos
+	}{
+		{
+			name:     "Use flat fee",
+			flatFee:  true,
+			fee:      1001,
+			expected: 1001,
+		},
+		{
+			name:     "Flat fee does not get overridden with min fee",
+			flatFee:  true,
+			fee:      999,
+			expected: 999,
+		},
+		{
+			name:     "Estimated fee overridden with min fee",
+			flatFee:  false,
+			fee:      1,
+			expected: 1000,
+		},
+		{
+			name:     "Flat fee set to 0",
+			flatFee:  true,
+			fee:      0,
+			expected: 0,
+		},
+	}
+	addr := types.Address{}.String()
+	for _, testcase := range testcases {
+		t.Run(testcase.name, func(t *testing.T) {
+			var tx types.Transaction
+			var err error
+
+			params := types.SuggestedParams{
+				FlatFee:         testcase.flatFee,
+				Fee:             testcase.fee,
+				FirstRoundValid: 1,
+				LastRoundValid:  1001,
+				GenesisHash:     byteFromBase64("JgsgCaCTqIaLeVhyL6XlRu3n7Rfk2FxMeK+wRSaQ7dI="),
+			}
+
+			tx, err = MakeAssetTransferTxn(addr, addr, 1, nil, params, "", 1)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetAcceptanceTxn(addr, nil, params, 1)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetRevocationTxn(addr, addr, 1, addr, nil, params, 1)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetDestroyTxn(addr, nil, params, 1)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetCreateTxn(addr, nil, params, 1, 1, false, addr, addr, addr, addr, "", "", "", "")
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetConfigTxn(addr, nil, params, 1, addr, addr, addr, addr, false)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+
+			tx, err = MakeAssetFreezeTxn(addr, nil, params, 1, addr, true)
+			require.NoError(t, err)
+			require.Equal(t, testcase.expected, tx.Fee)
+		})
+	}
+}
+
+func TestParseBoxReferences(t *testing.T) {
+
+	genWithAppId := func(appId uint64) types.AppBoxReference {
+		return types.AppBoxReference{appId, []byte("example")}
+	}
+
+	genWithNewAppId := func() types.AppBoxReference {
+		return types.AppBoxReference{0, []byte("example")}
+	}
+
+	t.Run("appIndexExists", func(t *testing.T) {
+		appId := uint64(7)
+		abr := genWithAppId(appId)
+
+		brs, err := parseBoxReferences(
+			[]types.AppBoxReference{abr},
+			[]uint64{1, 3, 4, appId},
+			appId-1)
+		require.NoError(t, err)
+		require.Equal(t,
+			[]types.BoxReference{{
+				ForeignAppIdx: uint64(4),
+				Name:          abr.Name}},
+			brs)
+	})
+
+	t.Run("appIndexDoesNotExist", func(t *testing.T) {
+		appId := uint64(7)
+		abr := genWithAppId(appId)
+
+		_, err := parseBoxReferences(
+			[]types.AppBoxReference{abr},
+			[]uint64{1, 3, 4},
+			appId-1)
+		require.Error(t, err)
+	})
+
+	t.Run("newAppId", func(t *testing.T) {
+		abr := genWithNewAppId()
+
+		brs, err := parseBoxReferences(
+			[]types.AppBoxReference{abr},
+			[]uint64{},
+			uint64(1))
+		require.NoError(t, err)
+		require.Equal(t,
+			[]types.BoxReference{{
+				ForeignAppIdx: uint64(0),
+				Name:          abr.Name}},
+			brs)
+	})
+
+	t.Run("fallbackToCurrentApp", func(t *testing.T) {
+		// Mirrors priority search in goal from `cmd/goal/application.go::translateBoxRefs`.
+		appId := uint64(7)
+		abr := genWithAppId(appId)
+
+		// Prefer foreign apps index when present.
+		brs, err := parseBoxReferences(
+			[]types.AppBoxReference{abr},
+			[]uint64{1, 3, 4, appId},
+			appId)
+		require.NoError(t, err)
+		require.Equal(t,
+			[]types.BoxReference{{
+				ForeignAppIdx: uint64(4),
+				Name:          abr.Name}},
+			brs)
+
+		// Fallback to current app when absent from foreign apps.
+		brs, err = parseBoxReferences(
+			[]types.AppBoxReference{abr},
+			[]uint64{1, 3, 4},
+			appId)
+		require.NoError(t, err)
+		require.Equal(t,
+			[]types.BoxReference{{
+				ForeignAppIdx: uint64(0),
+				Name:          abr.Name}},
+			brs)
+	})
 }
